Implemented cross origin solution through annotations in the controller
