+++ src/main/resources/alysopedia-app/.gitignore
