+++ src/main/resources/alysopedia-app/README.md
