+++ src/main/resources/alysopedia-app/public/logo192.png
