+++ src/main/resources/alysopedia-app/public/logo512.png
