+++ src/main/resources/alysopedia-app/public/manifest.json
