+++ src/main/resources/alysopedia-app/src/index.js
