+++ src/main/resources/alysopedia-app/src/serviceWorker.js
